+++ CMakeLists.txt
@@ -1,0 +1,7 @@
+cmake_minimum_required(VERSION 3.15)
+project(saleb LANGUAGES C VERSION 0.1.0)
+
+file(GLOB_RECURSE SALEB_SOURCES RELATIVE ${PROJECT_SOURCE_DIR} "src/*.c")
+file(GLOB_RECURSE SALEB_MAIN_SOURCE RELATIVE ${PROJECT_SOURCE_DIR} "test/saleb.c")
+
+add_executable(SALEB ${SALEB_MAIN_SOURCE} ${SALEB_SOURCES})
